Refactor robot file organization and specs
Spec updated to follow guidelines at: betterspecs.org
Files separated out
Robot verify returns struct rather than hash
